feat: modernize mvp_rxjava with RxJava 2, Room 2.6, ViewBinding and Firebase Firestore backend

+++ funnyvote/app/src/main/java/com/heaton/funnyvote/ui/WelcomeActivity.java
@@ -5,30 +5,23 @@ import android.content.Intent;
 import android.content.SharedPreferences;
 import android.os.AsyncTask;
 import android.os.Bundle;
-import android.support.annotation.Nullable;
-import android.support.v7.app.AppCompatActivity;
+
+import androidx.annotation.Nullable;
+import androidx.appcompat.app.AppCompatActivity;
 
 import com.heaton.funnyvote.FirstTimePref;
-import com.heaton.funnyvote.FunnyVoteApplication;
 import com.heaton.funnyvote.MainActivity;
 import com.heaton.funnyvote.R;
 import com.heaton.funnyvote.data.Injection;
+import com.heaton.funnyvote.data.local.AppDatabase;
+import com.heaton.funnyvote.data.local.dao.PromotionDao;
 import com.heaton.funnyvote.data.user.UserDataRepository;
 import com.heaton.funnyvote.database.Promotion;
-import com.heaton.funnyvote.database.PromotionDao;
-import com.heaton.funnyvote.database.User;
 import com.heaton.funnyvote.ui.introduction.IntroductionActivity;
 
 import java.lang.ref.WeakReference;
 import java.util.ArrayList;
 import java.util.List;
-
-import rx.Observer;
-import rx.subscriptions.CompositeSubscription;
-
-/**
- * Created by heaton on 2016/10/26.
- */
 
 public class WelcomeActivity extends AppCompatActivity {
 
@@ -40,24 +33,19 @@ public class WelcomeActivity extends AppCompatActivity {
         task.execute();
     }
 
-    @Override
-    protected void onDestroy() {
-        super.onDestroy();
-    }
-
     private static class InitTask extends AsyncTask<Void, Void, Void> {
 
         private WeakReference<Activity> contextWeakReference;
         private SharedPreferences firstTimePref;
 
         public InitTask(Activity context) {
-            contextWeakReference = new WeakReference<Activity>(context);
+            contextWeakReference = new WeakReference<>(context);
             firstTimePref = Injection.provideFirstTimePref(contextWeakReference.get());
         }
 
-
         @Override
         protected void onPostExecute(Void o) {
+            if (contextWeakReference.get() == null) return;
             if (firstTimePref.getBoolean(FirstTimePref.SP_FIRST_INTRODUCTION_PAGE, true)) {
                 contextWeakReference.get().startActivity(
                         new Intent(contextWeakReference.get(), IntroductionActivity.class));
@@ -66,52 +54,47 @@ public class WelcomeActivity extends AppCompatActivity {
                         new Intent(contextWeakReference.get(), MainActivity.class));
             }
             UserDataRepository userDataRepository = Injection.provideUserRepository(contextWeakReference.get());
-            CompositeSubscription subscription = new CompositeSubscription();
-            subscription.add(userDataRepository.getUser(true)
+            userDataRepository.getUser(true)
                     .subscribeOn(Injection.provideSchedulerProvider().computation())
                     .observeOn(Injection.provideSchedulerProvider().ui())
-                    .subscribe(new Observer<User>() {
-                        @Override
-                        public void onCompleted() {
-
-                        }
-
-                        @Override
-                        public void onError(Throwable e) {
-                            contextWeakReference.get().finish();
-                        }
-
-                        @Override
-                        public void onNext(User user) {
-                            contextWeakReference.get().finish();
-                        }
-                    }));
+                    .subscribe(
+                            user -> {
+                                if (contextWeakReference.get() != null) {
+                                    contextWeakReference.get().finish();
+                                }
+                            },
+                            throwable -> {
+                                if (contextWeakReference.get() != null) {
+                                    contextWeakReference.get().finish();
+                                }
+                            }
+                    );
         }
 
         @Override
         protected Void doInBackground(Void... voids) {
             try {
-                Thread.currentThread();
                 Thread.sleep(500);
             } catch (InterruptedException e) {
                 e.printStackTrace();
             }
-            //if (firstTimePref.getBoolean(FirstTimePref.SP_FIRST_MOCK_DATA, true)) {
-            PromotionDao promotionDao = ((FunnyVoteApplication) (contextWeakReference.get().getApplicationContext()))
-                    .getDaoSession().getPromotionDao();
-            String imageURL[] = contextWeakReference.get().getResources().getStringArray(R.array.imageURL);
-            List<Promotion> promotions = new ArrayList<>();
-            for (int i = 0; i < 1; i++) {
-                Promotion promotion = new Promotion();
-                promotion.setImageURL(imageURL[i % imageURL.length]);
-                promotion.setActionURL("https://play.google.com/store/apps/details?id=com.heaton.funnyvote");
-                promotion.setTitle("title:" + i);
-                promotions.add(promotion);
+            if (firstTimePref.getBoolean(FirstTimePref.SP_FIRST_MOCK_DATA, true)) {
+                if (contextWeakReference.get() != null) {
+                    PromotionDao promotionDao = AppDatabase.getInstance(contextWeakReference.get().getApplicationContext()).promotionDao();
+                    String[] imageURL = contextWeakReference.get().getResources().getStringArray(R.array.imageURL);
+                    List<Promotion> promotions = new ArrayList<>();
+                    for (int i = 0; i < 1; i++) {
+                        Promotion promotion = new Promotion();
+                        promotion.setImageURL(imageURL[i % imageURL.length]);
+                        promotion.setActionURL("https://play.google.com/store/apps/details?id=com.heaton.funnyvote");
+                        promotion.setTitle("title:" + i);
+                        promotions.add(promotion);
+                    }
+                    promotionDao.deleteAll();
+                    promotionDao.insertAll(promotions);
+                    firstTimePref.edit().putBoolean(FirstTimePref.SP_FIRST_MOCK_DATA, false).apply();
+                }
             }
-            promotionDao.deleteAll();
-            promotionDao.insertInTx(promotions);
-            firstTimePref.edit().putBoolean(FirstTimePref.SP_FIRST_MOCK_DATA, false).apply();
-            //}
             return null;
         }
     }
